Pull: extract recursive call to BindBind cont
The tailRecM - stack safety errors that we observed were genuine, and
they were caused by we having an unlimited chain of recursive calls
to the `cont` method from the BindBind class.

To prevent this, we extract a loop (a tail-recursive method) that
keeps unrolling this chain for as long as it is only terminals that
it meets throughout.