Pull Compilation - Remove BuildR (#2445)
Currently, in the recursive calls to `go`, we are using the `BuildR` 
runner as a way to control recursion. Instead, we can use the 
runners that are built at each step.

This "simplification", was tried before, but it was causing Memory Leaks.
We diagnosed those (with heapdump) as being caused by the compiler
creating many objects of type `BindBind`, or `DelegateBind`, in which
the  `step` (pull on the left hand side of >>=) was just a terminal, 
which is to say, a `Bind` that can be immediately reduced. 
To avoid this problem, we add a "smart constructor" method 
`bindView`, used to bind a Pull with a View.
This constructor, instead of always creating a new object, tries
to strictly apply the evaluation-reduction instead.

This "smart" construction introduced another problem, that the 
compilation of stream was not stack-safe, as a stack overflow 
could occur when evaluating a long chain of recursive calls
to the `cont` method of `BindBind` objects. 
To prevent this, we extract a loop (a tail-recursive method) that
keeps unrolling this chain for as long as it is only terminals that
it meets throughout.