Pull - Avoid View Objects, use mutable variable in compile
The `ViewL` trait, associated to the Pull GADT, has the purpose
of attaching an Action pull (a leaf in our tree) with the
continuation, usually given by a Bind object, on how the
tail of the stream is built from the result of the action.

- We define the trait `ContP`, for objects that include a function
  to continue from a Terminal Pull to a new Pull. This trait has
  two subclasses: Bind pulls, and an Identity.

- We add to the `compile` function a  mutable variable `contP` variable.
  We modify the `viewL` unroll-rotate loop, so that it stores the
  continuation sections (binds or identity) in this variable.
  Every call to `viewL` is immediately followed by reading this
  mutable variable.  Unfortunately, to use some `asInstanceOf`
  annotations when assigning and retrieving the `contP` field.

- We turn `Action` into an instance of the `ViewL` trait. Thus,
  the rotating/unrolling procedure `viewL` would now yield
  either a terminal or an action.

We think this could improve memory efficiency, since it completely
avoids any View objects. A relevant factor for Garbage Collection cost
is the number of  objects allocated in the heap. Fewer, fatter objects
is easier to collect than smaller ones.

Given that different calls to compile would be using different variables,
and if we ensure that the variable is read almost as soon as it is
written into (to avoid long windows), this should not risk interferences.

- build.sbt: we add Mima Filters for removed classes.