Pull - FlatMapR: Replace constructor with smart constructor.
Before, we were calling to constructor of FlatMapOutput, which
always created an object even though the tail may turn out to
be just a Terminal object, a Succeeded(Unit), which is to say
a nothing. So we got infinite chains for nothing.

The smart constructor matches if tail is a Terminal and, if so,
ignores the FlatMapOut and returns the terminal....