Pull - Translate - fix compilation bug.
Looking at the translate method, I noticed that, strangely enough,
it was not using the `view` or continuation. This means that any
stream after a Translate would be dropped.

This was not detected because, in the StreamTranslateSuite, none
of the unit/property tests was appending something after the
translation. This would be embarrassing.

Along the way, we also add a leak test that uses translate.