Pull Compilation Tweak: reuse BuildR object.
In the compilation loop, the BuildR simply represents a sort-of
callback or continuation, to connect the result of an inner stream.
We use them because directly passing the Uncons or the StepLeg runners
currently causes memory leaks.

However, since the BuildR object itself has no state, we can just
reuse a same object all through the compilation.