Drop-Right: reimplement to keep chunk structure
The current implementation of drop-right is constantly splitting
and appending chunks. This means that the "chunk structure", the
partitioning of output items in the chunks, is distorted by a
method that should not need do that.

This commit canges the implementation. Instead of Chunk-append
and split, we use a Queue of chunks and a total size counter.
Chunks from the source are added to this queue until total size
is enough to cover the number of dropped elements. Afterwards,
whenever a chunk can be dropped while keeping that covered,
the chunk is emitted as output.