Pull - Cut ties with BindBind, no need for the BuildR
We modify `Bind` class to make the `step` into an abstract method.

For the `BindBind` process, we modify the `bindBindAux` method
(for applying continuation) to cut ties to the two Bind objects.
An intuition for this is that a `BindBind` is just a node in a
stack of continuations, applied to the terminal result of a pull.
Once applied, the `BindBind` should not be used again.

Cutting this ties allows us to remove the use of the `BuildR`
intermediate representation, for the Uncons and StepLeg
case interpreters, without this causing the memory leaks tests
to fail.