Chunk - Introduced an "Interspersed" class.
The current intersperse method would, from a 1024 elements array,
generate a 2048 elements out of which half are just the separator.
This is not good. Since "Chunk" is just an interface, we can add
a bespoke Chunk implementation and thus change memory for CPU.

Note that the needed CPU is just a few bitwise / arithmetic ops.