Fix StackOverflow bug 3011.
Issue 3011 https://github.com/typelevel/fs2/issues/3011
showed an existing StackOverflow loop, caused by the
short-cut optimisation of `transformWith` together with
the recursion. This only happened when the FlatMap was
taking an extra large chunk (as was the case in example)

The solution is to add an internal whirl-loop, which
optimises this cyclic path away in tail-recursive way.