Document the 2.ms value
On my M2 Max running 1000 times in a row with a granularity of 1ms, I
get values ranging from 1 to 88 ms, average just under 18ms and std dev
13.6.

The longest sleep is a loop iteration 44 times, the shortest is 1. This
feels like 2ms is an okay value, especially considering this work is
most likely occurring after a server has served its useful purpose.